+++ list-zipper.nix
@@ -1,13 +1,14 @@
-{ mkDerivation, base, checkers, comonad, deriving-compat, lens
+{ mkDerivation, base, checkers, comonad, deriving-compat, lens, mtl
 , QuickCheck, semigroupoids, semigroups, stdenv, tasty, tasty-hunit
-, tasty-quickcheck
+, tasty-quickcheck, transformers
 }:
 mkDerivation {
   pname = "list-zipper";
-  version = "0.0.1";
+  version = "0.0.3";
   src = ./.;
   libraryHaskellDepends = [
-    base comonad deriving-compat lens semigroupoids semigroups
+    base comonad deriving-compat lens mtl semigroupoids semigroups
+    transformers
   ];
   testHaskellDepends = [
     base checkers lens QuickCheck tasty tasty-hunit tasty-quickcheck